Extract fake plan details creation functions so that they can be reused.
