+++ pkg/service/service_test.go
@@ -3,13 +3,14 @@ package service
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/redhat-developer/odo/pkg/testingutil"
+
 	"reflect"
 	"sort"
 	"testing"
 
 	scv1beta1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	appsv1 "github.com/openshift/api/apps/v1"
-	"github.com/pkg/errors"
 	applabels "github.com/redhat-developer/odo/pkg/application/labels"
 	componentlabels "github.com/redhat-developer/odo/pkg/component/labels"
 	"github.com/redhat-developer/odo/pkg/occlient"
@@ -18,87 +19,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	ktesting "k8s.io/client-go/testing"
 )
-
-// M is an alias for map[string]interface{}
-type M map[string]interface{}
-
-func fakePlanExternalMetaDataRaw() ([][]byte, error) {
-	planExternalMetaData1 := make(map[string]string)
-	planExternalMetaData1["displayName"] = "plan-name-1"
-
-	planExternalMetaData2 := make(map[string]string)
-	planExternalMetaData2["displayName"] = "plan-name-2"
-
-	planExternalMetaDataRaw1, err := json.Marshal(planExternalMetaData1)
-	if err != nil {
-		return nil, errors.Wrap(err, "")
-	}
-
-	planExternalMetaDataRaw2, err := json.Marshal(planExternalMetaData2)
-	if err != nil {
-		return nil, errors.Wrap(err, "")
-	}
-
-	var data [][]byte
-	data = append(data, planExternalMetaDataRaw1)
-	data = append(data, planExternalMetaDataRaw2)
-
-	return data, nil
-}
-
-func fakePlanServiceInstanceCreateParameterSchemasRaw() ([][]byte, error) {
-	planServiceInstanceCreateParameterSchema1 := make(M)
-	planServiceInstanceCreateParameterSchema1["required"] = []string{"PLAN_DATABASE_URI", "PLAN_DATABASE_USERNAME", "PLAN_DATABASE_PASSWORD"}
-	planServiceInstanceCreateParameterSchema1["properties"] = map[string]M{
-		"PLAN_DATABASE_URI": {
-			"default": "someuri",
-			"type":    "string",
-		},
-		"PLAN_DATABASE_USERNAME": {
-			"default": "name",
-			"type":    "string",
-		},
-		"PLAN_DATABASE_PASSWORD": {
-			"type": "string",
-		},
-		"SOME_OTHER": {
-			"default": "other",
-			"type":    "string",
-		},
-	}
-
-	planServiceInstanceCreateParameterSchema2 := make(M)
-	planServiceInstanceCreateParameterSchema2["required"] = []string{"PLAN_DATABASE_USERNAME_2", "PLAN_DATABASE_PASSWORD"}
-	planServiceInstanceCreateParameterSchema2["properties"] = map[string]M{
-		"PLAN_DATABASE_USERNAME_2": {
-			"default": "user2",
-			"type":    "string",
-		},
-		"PLAN_DATABASE_PASSWORD": {
-			"type": "string",
-		},
-	}
-
-	planServiceInstanceCreateParameterSchemaRaw1, err := json.Marshal(planServiceInstanceCreateParameterSchema1)
-	if err != nil {
-		if err != nil {
-			return nil, errors.Wrap(err, "")
-		}
-	}
-
-	planServiceInstanceCreateParameterSchemaRaw2, err := json.Marshal(planServiceInstanceCreateParameterSchema2)
-	if err != nil {
-		if err != nil {
-			return nil, errors.Wrap(err, "")
-		}
-	}
-
-	var data [][]byte
-	data = append(data, planServiceInstanceCreateParameterSchemaRaw1)
-	data = append(data, planServiceInstanceCreateParameterSchemaRaw2)
-
-	return data, nil
-}
 
 func TestGetServiceClassAndPlans(t *testing.T) {
 
@@ -112,13 +32,13 @@ func TestGetServiceClassAndPlans(t *testing.T) {
 		return
 	}
 
-	planExternalMetaDataRaw, err := fakePlanExternalMetaDataRaw()
+	planExternalMetaDataRaw, err := testingutil.FakePlanExternalMetaDataRaw()
 	if err != nil {
 		fmt.Printf("error occured %v during marshalling", err)
 		return
 	}
 
-	planServiceInstanceCreateParameterSchemasRaw, err := fakePlanServiceInstanceCreateParameterSchemasRaw()
+	planServiceInstanceCreateParameterSchemasRaw, err := testingutil.FakePlanServiceInstanceCreateParameterSchemasRaw()
 	if err != nil {
 		fmt.Printf("error occured %v during marshalling", err)
 		return
